requirements.txt
fastapi
uvicorn
requests
python-multipart
PyPDF2
faiss-cpu
sentence-transformers
google-generativeai
